% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftable.R
\name{fmmodel}
\alias{fmmodel}
\title{fmmodel}
\usage{
fmmodel(tab, ftsz = 9, ftname = "Time New Roma")
}
\arguments{
\item{tab}{A data frame}

\item{ftsz}{the size of font in the table}

\item{ftname}{the name of font in the table}
}
\value{
The default method returns table
}
\description{
the function is to format a model table style from html file  produced by stargazer package
}
\examples{
\dontrun{
fname<-system.file("extdata","table.html",package = "startowd",mustWork = TRUE)
table<-gettable(fname)
print(table)
ft<-fmmodel(table)
ft
}

}
