% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatdata.R
\name{fmtdata}
\alias{fmtdata}
\alias{fatdat}
\title{fmtdata}
\usage{
fatdat(
  x,
  langu = "C",
  tabname = "",
  notehead = "",
  ftsz = 9,
  ftname = "Time New Roma",
  align = "center"
)
}
\arguments{
\item{x}{A data frame}

\item{langu}{is a mark which means your paper enviroment is chinese or English, the default is "C", if you want to create a table in English, you can set parameter langu="E".}

\item{tabname}{add table name to header}

\item{notehead}{is table footer, the default is none.}

\item{ftsz}{the size of font in the table(default:9pt)}

\item{ftname}{the name of font in the table(default:Time New Roma)}

\item{align}{a align way of the table text(default:center)}
}
\value{
The default method returns table
}
\description{
the function is to format a summary data
}
\examples{
\dontrun{
fname<-system.file("extdata","summary.html",package = "startowd",mustWork = TRUE)
table<-gettable(fname)
ft<-fatdat(table)
ft
}


}
