% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftable.R
\name{fmmodel}
\alias{fmmodel}
\title{fmmodel}
\usage{
fmmodel(tab)
}
\arguments{
\item{tab}{A data frame}
}
\value{
The default method returns table
}
\description{
the function is to format a model table style from html file  produced by Stargazer package
}
\examples{
\dontrun{
fname<-system.file("extdata","table.html",package = "Startowd",mustWork = TRUE)
table<-gettable(fname)
print(table)
ft<-fmmodel(table)
ft
}

}
