% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftable.R
\name{fmmodel}
\alias{fmmodel}
\title{fmmodel}
\usage{
fmmodel(
  tab,
  langu = "C",
  tabname = "",
  notehead = "",
  notefoot = "",
  ftsz = 9,
  ftname = "Time New Roma"
)
}
\arguments{
\item{tab}{A data frame}

\item{langu}{is a mark which means your paper enviroment is chinese or English, the default is "C", if you want to create a table in English, you can set parameter langu="E".}

\item{tabname}{add table name to header}

\item{notehead}{is table footer, the default is none.}

\item{ftsz}{is a integer, you can set you table font size by it. the default is 9.}

\item{ftname}{is table font name.  the default is  "Time New Roma".}
}
\value{
The default method returns table
}
\description{
the function is to format a model table style from html file  produced by stargazer package
}
\examples{
\dontrun{
fname<-system.file("extdata","table.html",package = "startowd",mustWork = TRUE)
table<-gettable(fname)
print(table)
ft<-fmmodel(table)
ft
}

}
