% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatdata.R
\name{fmtdata}
\alias{fmtdata}
\alias{fatdat}
\title{fmtdata}
\usage{
fatdat(x, ftsz = 9, ftname = "Time New Roma", align = "center")
}
\arguments{
\item{x}{A data frame}

\item{ftsz}{the size of font in the table}

\item{ftname}{the name of font in the table(default:9pt)}

\item{align}{a align way of the table text(default:Time New Roma)}
}
\value{
The default method returns table
}
\description{
the function is to format a summary data
}
\examples{
\dontrun{
fname<-system.file("extdata","summary.html",package = "startowd",mustWork = TRUE)
table<-gettable(fname)
ft<-fatdat(table)
ft
}


}
