% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatdata.R
\name{fmtdata}
\alias{fmtdata}
\alias{fatdat}
\title{fmtdata}
\usage{
fatdat(x, align = "center")
}
\arguments{
\item{x}{A data frame}

\item{align}{a align way of table text}
}
\value{
The default method returns table
}
\description{
the function is to format a summary data
}
\examples{
\dontrun{
fname<-system.file("extdata","summary.html",package = "startowd",mustWork = TRUE)
table<-gettable(fname)
ft<-fatdat(table)
ft
}


}
